# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for USER Cukur-in",
        "title": "User Cukur-in",
        "contact": {
            "name": "Nuryanto",
            "url": "https://www.linkedin.com/in/nuryanto-1b2721156/",
            "email": "nuryantofattih@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/user-service/api/fileupload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FileUpload"
                ],
                "summary": "File Upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "account image",
                        "name": "upload_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "path images",
                        "name": "path",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
...
        "/user-service/user/capster": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Capster"
                ],
                "summary": "GetList Capster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Latitude",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perpage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "InitSearch",
                        "name": "initsearch",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortField",
                        "name": "sortfield",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModelList"
                        }
                    }
                }
            }
        },
        "/user-service/user/capster/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Capster"
                ],
                "summary": "GetById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Latitude",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
...
        "/user-service/user/feedback_rating": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FeedbackRating"
                ],
                "summary": "Add FeedbackRating",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddFeedbackRating"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user-service/user/feedback_rating/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FeedbackRating"
                ],
                "summary": "Rubah FeedbackRating",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddFeedbackRating"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AddFeedbackRating": {
            "type": "object",
            "properties": {
                "barber_id": {
                    "type": "integer"
                },
                "barber_rating": {
                    "type": "number"
                },
                "capster_id": {
                    "type": "integer"
                },
                "capster_rating": {
                    "type": "number"
                },
                "comment": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ResponseModelList": {
            "type": "object",
            "properties": {
                "all_column": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "define_column": {
                    "type": "string"
                },
                "define_size": {
                    "type": "string"
                },
                "last_page": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "tool.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "description": "Code int         `json:\"code\"`",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}